Try creating a config_local.py in Travis
The config updates we make in conftest.py don't seem to make it to flask_mail, so it still tries to send emails on Travis.
So we're now trying to set Testing=True before the app is created. This should also remove the need for setting the send_register_email to false and setting mail_suppress_send to True.