Temporarily replace cmenergies aggregation with dummy facet filter
ES 7.1 (current version on CERN servers) doesn't support histogram aggregations on range datatypes, so we need to remove the aggregations until we can upgrade to ES7.4.
We don’t want to lose the aggregations work by leaving it on a stale branch, so instead we’ve commented it out.
The aggregation has been replaced by a dummy facet filter which does not include doc counts but allows users to easily select a range of cmenergies values.