Use celery to schedule processing of uploads to avoid timeouts
When a user uploads a file, the file is now simply saved and a message shown to the user. The file is then processed by celery, and the user is emailed with success or failure notifications.

Added a simple test to records_test but this should possibly be expanded on in future or switched to an e2e test.