Use custom db field for bytea which are really (unicode) strings
As python3 distinguishes between text and binary data, we should treat
text as strings (which now handle unicode), but we don't want to change
the database. So for fields which are really strings stored as bytes,
I've created a LargeBinaryString type which does the conversion as the
data goes into and comes out of the database. This removes the need to
encode and decode strings elsewhere in the python code.

Needed to modify the yaml loader to return unicode strings in all cases,
rather than bytestrings.

Also replaced get_prefilled_dict with get_license as that was its only
purpose.