Got docker working again
 * The inspire python3.6 image is based on a Debian image rather than Centos,
and does not install npm. To get node-sass to install successfully on
Debian I've followed the instructions at
https://github.com/nodesource/distributions/blob/master/README.md#debinstall
as recommended by node-sass. Also updated the version of node-sass installed.
 * Added a step to the Dockerfile to install the `tests` dependencies on a
local web build, as they are no longer in requirements.txt
 * Use 'local-web' as the required value for APP_ENVIRONMENT to ease my
confusion when the build happened again for the worker node (which
doesn't need the tests)
 * Updated installation instructions to reflect my discoveries about node-sass