Remove attempt to create hepsubmission if it doesn't exist
The code was presumably never reached previously, as it could never have
worked, as it tried to access variables of the null submission. We don't
have enough information at this point to sensibly create a hepsubmission,
so let's not try. All calls to process_submission_directory rely on the
submission already being in existence, so we shouldn't get to an error
case here.