Sanitize HTML in submission abstract and table descriptions
We want to escape rather than strip disallowed tags so we use bleach
rather than invenio-formatter.
We have to take a guess about how to deal with text that is not HTML but
contains < or > - this works for the test cases found but we should add
tips in the hepdata-submission docs for what to do if the HTML output is
not as expected.