Fixed issue where sandbox data files from .oldhepdata were being deleted
cleanup_old_files had a shortcut for standard sandbox entries where the
current filepath was passed in. Unfortunately this didn't work for
oldhepdata uploads as the current_filepath pass in was the oldhepdata dir
not the converted yaml dir, so the newly created yaml files were deleted.

Decided to just remove the shortcut and use the safer option of getting
data file paths from the db.