Reduce risk of tests hanging  when fetching resources from hepdata.cedar.ac.uk
Added a timeout when fetching resources from old hepdata site. Used the
config NGINX_TIMEOUT for consistency with other url timeouts. Timeout
error is caught in except block in submission.py. Note that this was only
used for migration so shouldn't be needed in production.

Also removed most resources with /resource in the path in the test
submission to reduce chances of a timeout occurring in the tests.